pyproject.toml
[project]
name = "base-pipeline"
version = "0.1.0"
description = "Pipeline de visão computacional com PyTorch (auto GPU/CPU)"
authors = [{ name = "VisionLab" }]
requires-python = ">=3.12,<3.13"

dependencies = [
    "opencv-python",
    "scikit-image",
    "scikit-learn",
    "numpy",
    "ultralytics",
    "pandas",
    "matplotlib",
    "seaborn",
    "tqdm",
    "albumentations",
    "Pillow",
    "tensorboard",
    "torchmetrics",
    "pyyaml",
    "hydra-core",
    "onnx",
    "onnxruntime",
    "torchinfo",
    "jupyterlab",
    "roboflow>=1.2.11",
    "psutil",
]

[tool.uv]
managed = true
index-url = "https://pypi.org/simple"
index-strategy = "unsafe-best-match"
